LayoutTests/js/dom/modules/module-and-dom-content-loaded-expected.txt

Test window DOMContentLoaded event is fired after module is executed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Module is not executed yet.
PASS document.readyState is "loading"
PASS document.readyState is "interactive"
PASS domContentLoadedIsExecuted is false
PASS successfullyParsed is true

TEST COMPLETE
